chore(clerk-js): Extract user fixture creation to core
This will be used in core tests
